Small style fixes

+++ uAAL.utils/src/main/java/org/universAAL/support/utils/ui/SelectMulti.java
@@ -134,7 +134,7 @@ public class SelectMulti extends InputControl{
      * @return The index of the initial value
      */
     public Integer getInitialIndex() {
-        return initialValue.intValue();
+        return initialValue;
     }
 
     /**
